Add basic project Makefile
This has been late coming because I had been writing a good deal of code
to implement the pipeline in Make. For various reasons, I think we are
moving away from that idea, towards an interactive pipeline runner. This
is all of the make functionality we need for project management (setting
up an environment, linting, running tests).

There's one thing in here that may look a little strange: the
$(INSTALL_SENTINEL) line. What this is doing is touching a file whenever
poetry install was run successfully. Whenever poetry.lock has a newer
update time than that file (or if that file does not exist), then
`poetry install` will be run for any recipes that declare $(PYTHON) as a
dependency. In practice, that means you do not have to worry about
manually creating a virtual environment when running other targets (like
`make test` or `make lint`, which require `pytest` and `ruff` to be
installed, respectively).