Make some handy additions to the data validation targets
This is a followup to the changes in 1fbf3a8.

The big change here is to change the structure of log files created by
`linkml validate`, and to add some symlinks to more easily keep track of
those logs.  There's a lot of documentation added that explains the
logic behind what's going on, but briefly:

When a data file is validated, the log of that validation goes to a
timestamped file in a directory specific to that input file. If
validation was successful, that log file is symlinked to a `success.log`
file in the same directory. (That `success.log` is the target for the
make recipe, so it will not be revalidated in the future unless the data
file or schema changes).

If validation was not successful, then the log file is symlinked to
`latest-error.log` in the same directory. The directory is also
symlinked to a folder called `errors`, to give a list of all files which
did not pass validation.

If all input files were successfully validated, a sentinel file called
(`_validation_complete`) is touched. This file should be a prerequisite
for any further steps in the pipeline, as its existence marks that
validation succeeded.

---

Other changes: lots of documentation, some helpful logging